--- src/main/java/com/heydoctor/heydoctor/controller/mainpage/MainPageController.java
@@ -1,18 +1,0 @@
-package com.heydoctor.heydoctor.controller.mainpage;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@Slf4j
-@RequiredArgsConstructor
-@RequestMapping("/mainPage/*")
-public class MainPageController {
-    private final MainPageController mainPageController;
-    @GetMapping("mainPage")
-    public void mainPage(){ log.info("mainPage test");
-    }
-}
