+++ src/main/java/com/heydoctor/heydoctor/domain/service/adminpage/AdminpageService.java
@@ -1,0 +1,4 @@
+package com.heydoctor.heydoctor.domain.service.adminpage;
+
+public interface AdminpageService {
+}
