--- src/main/java/com/heydoctor/heydoctor/domain/service/impl/MypageImpl.java
@@ -1,10 +1,0 @@
-package com.heydoctor.heydoctor.domain.service.impl;
-
-import com.heydoctor.heydoctor.domain.service.inter.MypageService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MypageImpl implements MypageService {
-}
