--- src/main/java/com/heydoctor/heydoctor/domain/service/inter/AdminpageService.java
@@ -1,4 +1,0 @@
-package com.heydoctor.heydoctor.domain.service.inter;
-
-public interface AdminpageService {
-}
