--- src/main/java/com/heydoctor/heydoctor/domain/service/inter/MeetDoctorService.java
@@ -1,5 +1,0 @@
-package com.heydoctor.heydoctor.domain.service.inter;
-
-
-public interface MeetDoctorService {
-}
