--- src/main/java/com/heydoctor/heydoctor/domain/service/inter/QuestionBoardService.java
@@ -1,4 +1,0 @@
-package com.heydoctor.heydoctor.domain.service.inter;
-
-public interface QuestionBoardService {
-}
