+++ src/main/java/com/heydoctor/heydoctor/domain/service/loginpage/LoginPageService.java
@@ -1,0 +1,4 @@
+package com.heydoctor.heydoctor.domain.service.loginpage;
+
+public interface LoginPageService {
+}
