+++ src/main/java/com/heydoctor/heydoctor/domain/service/mainpage/MainPageService.java
@@ -1,0 +1,4 @@
+package com.heydoctor.heydoctor.domain.service.mainpage;
+
+public interface MainPageService {
+}
