+++ src/main/java/com/heydoctor/heydoctor/domain/service/meetdoctor/MeetDoctorService.java
@@ -1,0 +1,5 @@
+package com.heydoctor.heydoctor.domain.service.meetdoctor;
+
+
+public interface MeetDoctorService {
+}
