+++ src/main/java/com/heydoctor/heydoctor/domain/service/mypage/MyPageService.java
@@ -1,0 +1,4 @@
+package com.heydoctor.heydoctor.domain.service.mypage;
+
+public interface MyPageService {
+}
