+++ src/main/java/com/heydoctor/heydoctor/domain/service/mypage/MyPageServiceImpl.java
@@ -1,0 +1,10 @@
+package com.heydoctor.heydoctor.domain.service.mypage;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MyPageServiceImpl implements MyPageService {
+}
