+++ src/main/java/com/heydoctor/heydoctor/domain/service/questionboard/QuestionBoardService.java
@@ -1,0 +1,4 @@
+package com.heydoctor.heydoctor.domain.service.questionboard;
+
+public interface QuestionBoardService {
+}
